05-sincronització fet a classe

+++ 05-Sincronitzacio/src/Compte.java
@@ -16,7 +16,7 @@ public class Compte {
     public float getSaldo() {
         return saldo;
     }
-    public void setSaldo(int saldo) {
+    public void setSaldo(float saldo) {
         this.saldo = saldo;
     }
